requirements.txt
Flask==2.2.2
flask_smorest==0.40.0
geopy==2.3.0
python-dotenv==0.21.1
requests==2.28.1
